Moved kafka client
