+++ client/kafka/builder.go
