+++ client/kafka/builder_test.go
