+++ client/kafka/kafka_test.go
